Separated CQL ser/deser code from driver logic
Another crate was created in order to allow reuse of low-level code
in other projects (including under-progress proxy). The whole `frame`
module was made separate, along with structs regarding transport
errors. Some tests that require driver transport logic as dependency
were moved from `frame` to `transport`.
