+++ scylla-cql/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "scylla-cql"
+version = "0.0.1"
+edition = "2021"
+description = "CQL data types and primitives, for interacting with Scylla."
+repository = "https://github.com/scylladb/scylla-rust-driver"
+readme = "../README.md"
+keywords = ["database", "scylla", "cql", "cassandra"]
+categories = ["database"]
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+scylla-macros = { version = "0.1.1", path = "../scylla-macros"}
+byteorder = "1.3.4"
+bytes = "1.0.1"
+num_enum = "0.5"
+tokio = "1.12"
+snap = "1.0"
+uuid = "1.0"
+thiserror = "1.0"
+bigdecimal = "0.2.0"
+num-bigint = "0.3"
+chrono = "0.4"
+lz4_flex = { version = "0.9.2" }
