+++ scylla-cql/src/frame/frame_errors.rs
