+++ scylla-cql/src/frame/request/register.rs
