+++ scylla-cql/src/frame/request/startup.rs
