+++ scylla-cql/src/frame/response/authenticate.rs
