+++ scylla-cql/src/frame/response/event.rs
