+++ scylla-cql/src/frame/response/result.rs
