+++ scylla-cql/src/frame/response/supported.rs
