+++ scylla-cql/src/frame/server_event_type.rs
