+++ scylla-cql/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod errors;
+pub mod frame;
+#[macro_use]
+pub mod macros;
+
+pub use crate::frame::response::cql_to_rust;
+pub use crate::frame::response::cql_to_rust::FromRow;
+
+pub use crate::frame::types::Consistency;
