+++ scylla-cql/src/macros.rs
