+++ scylla/Cargo.toml
@@ -15,6 +15,7 @@ ssl = ["tokio-openssl", "openssl"]
 
 [dependencies]
 scylla-macros = { version = "0.1.1", path = "../scylla-macros"}
+scylla-cql = { version = "0.0.1", path = "../scylla-cql"}
 byteorder = "1.3.4"
 bytes = "1.0.1"
 futures = "0.3.6"
